requirements.txt
python-telegram-bot
python-dotenv
orjson
tweepy==4.13.0